linters errors corrected
